自选股
自选股

+++ Code/EasyFrame/EasyFrame/Network/EFEntity.h
@@ -23,8 +23,8 @@
 /*!
  *  @brief  快捷参数列表
  */
-@property (nonatomic,strong)  id dict;
+@property (nonatomic,strong)    id dict;
 
-@property(nonatomic,assign)int tag;
+@property(nonatomic,assign)     NSInteger tag;
 
 @end
